requirement.txt
python-dotenv
pandas
numpy
setuptools
pymongo
certifi
pymongo[srv]==3.12
pymongo

## -e .